kleaf: use %{label} for progress message (hermetic_tools)
Mechanical change for aesthetic reasons. This ensures we have a
consistent way of encoding labels in progress messages, which
facilitates profile parsing.

For hermetic_tools, we no longer use the label of the outer target
in progress message; just take the easy way out and use %{label}
for the internal target.

Bug: 344672600
Change-Id: I2174f4f2dcd90cd5d07f16c571939f58aa13ae2b