Revert "Kleaf: Remove non-existent attr from kernel_prebuilt_repo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ecc11fb-ed84-4c70-ba2c-01f0fd4e0d15). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 354016054
Change-Id: I91bf19d7f6c51d99e71d5a25f7a0994252375d1b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com