kleaf: Use platform-agnostic labels for build-tools
For dependencies on //prebuilts/build-tools, use the platform-agnostic
labels instead of inferring to a specific target. That way, we have a
central place (//build/kernel:hermetic-tools) to control the set of
binaries we want to use by changing the transition.

The same will be done for //prebuilts/kernel-build-tools, just not in
this change.

Bug: 325514706
Change-Id: I78ca229125d81c91a62b9b834feb8750a5e222f1