kleaf: hermetic_tools replace cfg=exec with hand-crafted transition
Explicitly transition --platforms to @platforms//host instead of using
cfg="exec". This allows us to replace the platform if a flag is set.

This also allows us to simplify hermetic_tools implementation.
Previously, we need a wrapper to do the cfg="exec" transition, because
cfg="exec" can't be applied on the incoming edge. Now with an explicit
transition to @platforms//host, we can apply it on the incoming edge,
simplifying things.

Bug: 325514706
Change-Id: Ic4801acb5ac70c12c4cef51784d4b2a01aa78838