kleaf: add --musl_prebuilts
When the flag is set, use musl variant for prebuilt binaries in as many
places as possible.

Prebuilt libraries are in a gray area; for prebuilt binaries, use
musl variant if --musl_prebuilts is set. Binaires built with
cc_ rules and Kbuild link to the glibc or musl variant based on their
corresponding flags, and are unaffected by --musl_prebuilts.

Bug: 325514706
Change-Id: Ieb8f712699cd1fb571334388e3fbc16df3ed61d7