kleaf: get rid of LD_LIBRARY_PATH
Based on the assumption that all host binaries built by Kbuild is one
level below $OUT_DIR, this change adds to HOSTLDFLAGS a list of
runpaths, so that these host binaries built by Kbuild may link to
prebuilts/kernel-build-tools at runtime.

This is a prerequisite to separate library search paths for Kbuild-built
host binaries (which is controlled by --musl_kbuild) and prebuilt host
binaries (which is controlled by --musl_prebuilts).

This also ensures the library search paths for //prebuilts/build-tools
and //prebuilts/kernel-build-tools are different (respects their
runpaths), so that when one is updated without the other, there won't be
compatibility bugs any more.

Caveat:

The only assumption of this change is that all host binaries built by
Kbuild (or at least the ones that needs to link against libraries below
//prebuilt/kernel-build-tools) is one level below $OUT_DIR,
e.g. certs/extract-cert, scripts/sign-file, etc. If this changes in the
future, kernel_env needs to add additional HOSTLDFLAGS to backtrack
more levels.

Bug: 206118938
Change-Id: I8de11da2dde8f365334cbabdad90b0c93ffe4087