kleaf: hermetic tools drop //prebuilts/build-tools:linux-x86
and //prebuilts/kernel-build-tools:linux-x86.

The explicit dependency is no longer needed; the libs comes from
the runfiles of individual tool from //prebuilts/build-tools and
//prebuilts/kernel-build-tools.

Implementation notes:
As a side-effect of this change, notably, the following files are
removed from hermetic tools dependency:
- prebuilts/build-tools/common/framework
- prebuilts/build-tools/common/flex/FlexLexer.h
- prebuilts/build-tools/common/py3-stdlib
- prebuilts/kernel-build-tools/*/share/avb/testkey_rsa2048.pem
- ... and other unnecessary binaires below **/bin/.

These files are deemed unnecessary during the kernel build, so we just
take them out to minimize the input list.

Bug: 325514706
Change-Id: I7448391a56ca2c7ed5639d4dc9c11673372ad56f