kleaf: --config=musl enables --host_platform=<musl>.
Now that everything in the canary build has switched to be able to build
against musl host platform, switch it over.

This change currently has the following effect:
- The python toolchian itself links against host musl libc
- The clang toolchain builds binaries against host musl libc by default,
  unless the binary has an explicit target_compatible_with clause.

Update cc_testing so it can be built with --host_platform=musl.

With `--host_platform=<musl>`, fully_static_link cc_binary's can be
built again. In fact, `-static` is hard-coded if linkstatic=True (which
is the default). Update cc_testing to reflect this.

Bug: 325514706
Bug: 348087663
Change-Id: I2824047094cfcfc1dcd8e583689e0f974758e26d