[Do Not Merge] Updated the build script to handle the page_size change
When a config is generated for 4K page and then again re-generated for
16K page, the value CONFIG_ARM64_VA_BITS_39 doesn't get unset, Which
after sorting the config, put this value after CONFIG_ARM64_VA_BITS_36
causing the build system to use 47 bits instead of 36 for 16K page size.
This change will unblock the 16k page config setup. Right now, this
seems to be the easiest solution.
This updates the CONFIG_ARM64_VA_BITS_<VA_BITS> value to "is not set" if
set to "y", for all the VA_BITS that are not equal to
CONFIG_ARM64_VA_BITS value.

Bug: 358270050
Test: make -c build kernel-menuconfig HW=postsil
This will create the config which will not have any duplicate
CONFIG_ARM64_VA_BITS_<VA_BITS> entries.
Change-Id: I26b8d9bcefd80510edc517a064a210f32c042c62
Signed-off-by: Tarun Sahu <tarunsahu@google.com>