build_utils.sh: Allow kunit modules to be included in device builds
Since TRIM_UNUSED_MODULES uses the modules.order file to determine which
modules are present, we need to retain the original modules.order file
from the GKI build in order to allow device kernels to use the kunit
modules. Update create_modules_staging() to backup and restore the
original modules.order file if TRIM_UNUSED_MODULES is not set.

Bug: 323942195
Change-Id: Ibefb2805b500c3fb944293e79b7b1b5301abcf58
Signed-off-by: Will McVicker <willmcvicker@google.com>