Revert "build_utils.sh: Allow kunit modules to be included in de..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/28e1c89f-983f-4dad-b22e-678c86fecad9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361527046
Change-Id: I9908975715b441923de8b40de63d949b609c9334
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com