kleaf: bison_wrapper uses runfiles.
Relies on runfiles_dir to find the runfiles when the executable is
executed at `bazel run`, e.g. when `bazel run xxx_config`. This is the
recommended way.

With this change, `bazel run //build/kernel:bison` works, though this is
not a public API and is not how bison is supposed to be invoked.

Bug: 325514706
Change-Id: I8230ac1d7bbf42262d7ebd674139acd66c7e57e2