kleaf: simplify swig_wrapper.bzl.
Don't use short_paths every where. We don't rely on runfiles dir, but
rather on $0, to find execroot. The paths below execroot is always
path, not short_path.

https://bazel.build/rules/lib/builtins/File#short_path says:
"Use the short_path for the path under which the file is mapped if it's
in the runfiles of a binary."

With this change, `bazel run //build/kernel:bison` works, though this is
not a public API and is not how bison is supposed to be invoked.

Bug: 325514706
Change-Id: Ife8feade309a2d405a4bf97db207d8b1c82b7acb