download_from_ci: Fix artifacts for virtual i686 kernel
The artifacts for kernel target kernel_virt_i686 has bzImage as the
kernel binary. Absense of its support resulted in corrupt kernel bins
being copied to the cuttlefish_prebuilts directory.

Fix the script so that bzImage is considered for this targe.

Bug: 362204914
Test: Executed `download_from_ci -u -r -b 353916950 12109370` and
verified that the kernel-6.1 bin in cuttlefish_prebuilts was proper.

Change-Id: I01319721740c0324ec6ff79e4c4dc428ebecadbc