kleaf: Merge linux_includes from submodules
Even though ddk_submodule is not allowed to specify linux_includes,
its dependency may specify so. In particular, specifying
deps=["//common:all_headers"] in ddk_submodule is the only way to get
sumodules build correctly prior to this change.

However, when the number of submodules grows, the items in LINUXINCLUDES
becomes repetitive because we handled LINUXINCLUDES in each submodule
then merge the Kbuild file. The generated file looks like this:

    # Common flags for submodule_0.o
    LINUXINCLUDE := ...

    # Common flags for submodule_1.o
    LINUXINCLUDE := ...

This change does not fix the issue immediately. Rather, it gathers all
LINUXINCLUDE for all submodules in the current directory, then emit a
single definition, with duplicates removed:

    # Common LINUXINLUDE for all submodules in this directory
    LINUXINCLUDE := ...

The LINUXINCLUDE from individual submodules are removed in follow up
changes.

Bug: 363143263
Change-Id: I862d0c166868a427c3cb6cf5aa208b7117ba47b1