Revert "Kleaf: experimentally enable pigz as drop-in-replacement..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41b8a4b8-6a8d-4550-919a-c9f1c1457f5f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice06cd47ae4d1869fc84c1c5d5e6846fde5b1aa8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com