kleaf: system_dlkm move fs_types calc to kernel_images.
This change deletes the deprecated system_dlkm_image.fs_type, making
the API cleaner, with the following cost:

Move fs_type calculation to loading phase, making
kernel_images.system_dlkm_fs_type[s] non-configurable. However, we
are going to deprecate kernel_images altogether soon, so this is
acceptable.

A side effect of this change is that, on the API level of
system_dlkm_image, it is no longer possible to build just
system_dlkm.img with the ext4 format; one always builds
system_dlkm.ext4.img. This is intentional to deprecate the fs_type attr.

Bug: 368689504
Change-Id: I4332e7d8034c6ef85ffeae2f23fb7e4bdbf5e24d