kleaf: kernel_config: Allow absolute path to defconfig
Kconfig support doesn't support an absolute path to a defconfig. The
defconfig must be relative to the srctree. In DDK, vendor/OEM wants to
maintain their own independent defconfig (IOW: independent of
gki_defconfig), thus the base defconfig cannot/does not live inside the
srctree. To support installing such a defconfig, we can copy the
defconfig file to the ${OUT_DIR}/.config and run make olddefconfig to
achieve a similar effect.

Bug: 290991919
Change-Id: I3e787c9f0c88e4c4a96f31792d660d06deaa26ca
Signed-off-by: Elliot Berman <quic_eberman@quicinc.com>