download_from_ci: Add --update-gki-signed (-s) option
This option updates the kernel prebuilts and virtual devices modules
in the "signed/" sub-directory.

NOTE: The script makes assumptions about the out directories already
exisiting and containing old kernel artifacts. This is not addressed
in this patch as it is an existing general issue and should be
addressed separately;  we are simply reusing update_gki().

The initial commits for signed builds are created at:

    https://r.android.com/q/topic:initial-gki-signed-prebuilts

Test: download_from_ci -s <build-id> -b <bug-id>
Bug: 372294660
Change-Id: I1fb6facb319a492ef67b7aef64a4622188272f9a