kleaf: Only build subgoals if there are any
Currently, the action in _kernel_build() will build the dtbs target
first if it is part of the make_goals attribute. It will then remove the
dtbs target from the make_goals and blindly build the remaining make
goals, even if there aren't any.

This is undesirable in cases where the only make goal is just the dtbs
target, as it unintentionally triggers a full kernel compilation.
Therefore, proceed to build the rest of the make goals if there are any.

Bug: 372546095
Bug: 361783887
Change-Id: If43086422a06bb5323cb8f598af9033935a0670a
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>