kleaf: avoid ROOT_DIR in HOSTRUSTFLAGS.
The flag added to runpath is embedded in the binary. If ROOT_DIR
is used, an absolute path is embedded in the binary, making it
non-reproducible and non-remote-cacheable. I apply the same trick
I did with HOSTLDFLAGS to HOSTRUSTFLAGS to find libunwind.so
properly.

For now, scripts/generate_rust_targets is the only instance. If we
have more binaries later, additional levels may be added.

Bug: 373703502
Test: bazel build //common:kernel_x86_64 --rust --nomusl_kbuild
Change-Id: Ibc48093c51345161cf190f11f9c4f04156536897